requirements.txt
aiogram==3.20.0.post0
PyMuPDF==1.24.0
google-api-python-client==2.170.0
google-auth-oauthlib==1.2.2
gspread==6.2.1
meteostat==1.6.8
mysql-connector-python==8.4.0
numpy==2.3.0
openai==1.86.0
openpyxl==3.1.2
pandas==2.3.0
Pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
pydicom==3.0.1
PyMySQL==1.1.0
python-dotenv==1.1.0
python-telegram-bot==22.1
python-dateutil==2.8.2
requests==2.32.4
schedule==1.2.2
sqlalchemy==2.0.41
